refactored stringutils to package for further use, and added more generic helpers to newer utils for ease of use
