+++ src/GradDataWarehousing/HW2/HW2.java
@@ -15,7 +15,7 @@ import java.util.concurrent.ConcurrentMap;
 import java.util.concurrent.atomic.AtomicInteger;
 
 import static Utils.ConsolePrinting.*;
-import static Utils.StringUtils.*;
+import static Utils.StringUtils.StringUtils.*;
 
 public class HW2 {
